Don't omit parts for ways that are also interiors
When a building has an exterior part and an interior part, write
the interior part (with tags) before the exterior part. Then when
we get to the exterior part, we can re-use the interior way as
role=inner in the more complicated relation.

FIXME: would fail on some complex cases. Seems to need two-pass
logic